[API] replace campaign to join repository
