+++ api/lib/domain/usecases/get-campaign-by-code.js
@@ -1,6 +1,11 @@
-module.exports = function getCampaignByCode({
+function getCampaignByCode({
   code,
   campaignToJoinRepository,
 }) {
   return campaignToJoinRepository.getByCode(code);
 };
+
+module.exports = {
+  perform: getCampaignByCode,
+  useTransaction: false,
+}
