+++ api/lib/domain/usecases/index.js
@@ -121,7 +121,6 @@ const dependencies = {
   userTutorialRepository: require('../../infrastructure/repositories/user-tutorial-repository'),
   verifyCertificateCodeService: require('../../domain/services/verify-certificate-code-service'),
 
-  campaignToJoinRepositoryTrx: require('../../infrastructure/repositories/campaign-to-join-repository-trx'),
   campaignParticipationRepositoryTrx: require('../../infrastructure/repositories/campaign-participation-repository-trx'),
   assessmentRepositoryTrx: require('../../infrastructure/repositories/assessment-repository-trx'),
 };
