+++ api/lib/infrastructure/repositories/campaign-to-join-repository.js
@@ -1,14 +1,14 @@
-const { knex } = require('../bookshelf');
 const CampaignToJoin = require('../../domain/read-models/CampaignToJoin');
 const { NotFoundError, ForbiddenAccess, AlreadyExistingCampaignParticipationError } = require('../../domain/errors');
 const skillDatasource = require('../datasources/learning-content/skill-datasource');
-const DomainTransaction = require('../DomainTransaction');
 
-module.exports = {
+class CampaignToJoinRepository {
+  constructor(queryBuilder) {
+    this.queryBuilder = queryBuilder;
+  }
 
-  async get(id, domainTransaction = DomainTransaction.emptyTransaction()) {
-    const knexConn = domainTransaction.knexTransaction || knex;
-    const result = await knexConn('campaigns')
+  async get(id) {
+    const result = await this.queryBuilder('campaigns')
       .select('campaigns.*')
       .select({
         'organizationId': 'organizations.id',
@@ -29,10 +29,15 @@ module.exports = {
     }
 
     return new CampaignToJoin(result);
-  },
+  }
+
+  async checkCampaignIsJoinableByUser(campaign, userId) {
+    await _checkCanAccessToCampaign(this.queryBuilder, campaign, userId);
+    await _checkCanParticipateToCampaign(this.queryBuilder, campaign, userId);
+  }
 
   async getByCode(code) {
-    const result = await knex('campaigns')
+    const result = await this.queryBuilder('campaigns')
       .select('campaigns.*')
       .select({
         'organizationId': 'organizations.id',
@@ -44,8 +49,7 @@ module.exports = {
         'targetProfileImageUrl': 'target-profiles.imageUrl',
         'targetProfileIsSimplifiedAccess': 'target-profiles.isSimplifiedAccess',
       })
-      // eslint-disable-next-line no-restricted-syntax
-      .select(knex.raw(`EXISTS(SELECT true FROM "organization-tags"
+      .select(this.queryBuilder.raw(`EXISTS(SELECT true FROM "organization-tags"
         JOIN tags ON "organization-tags"."tagId" = "tags".id
         WHERE "tags"."name" = 'POLE EMPLOI' AND "organization-tags"."organizationId" = "organizations".id) as "organizationIsPoleEmploi"`))
       .join('organizations', 'organizations.id', 'campaigns.organizationId')
@@ -60,65 +64,54 @@ module.exports = {
     }
 
     return new CampaignToJoin(result);
-  },
+  }
+}
 
-  async checkCampaignIsJoinableByUser(campaign, userId, domainTransaction = DomainTransaction.emptyTransaction()) {
-    await _checkCanAccessToCampaign(campaign, userId, domainTransaction);
-    await _checkCanParticipateToCampaign(campaign, userId, domainTransaction);
-  },
-};
-
-async function _checkCanAccessToCampaign(campaign, userId, domainTransaction) {
+async function _checkCanAccessToCampaign(queryBuilder, campaign, userId) {
   if (campaign.isArchived) {
     throw new ForbiddenAccess('Vous n\'êtes pas autorisé à rejoindre la campagne');
 
   }
 
-  if (campaign.isRestricted && await _hasNoActiveSchoolingRegistration(userId, campaign, domainTransaction)) {
+  if (campaign.isRestricted && await _hasNoSchoolingRegistration(queryBuilder, userId, campaign)) {
     throw new ForbiddenAccess('Vous n\'êtes pas autorisé à rejoindre la campagne');
   }
 }
 
-async function _hasNoActiveSchoolingRegistration(userId, campaign, domainTransaction) {
-  const knexConn = domainTransaction.knexTransaction || knex;
-  const registrations = await knexConn
-    .select('schooling-registrations.id')
+async function _hasNoSchoolingRegistration(queryBuilder, userId, campaign) {
+  const registrations = await queryBuilder .select('schooling-registrations.id')
     .from('schooling-registrations')
-    .where({ userId, organizationId: campaign.organizationId, isDisabled: false });
+    .where({ userId, organizationId: campaign.organizationId });
 
   return registrations.length === 0;
 }
 
-async function _checkCanParticipateToCampaign(campaign, userId, domainTransaction) {
-  if (campaign.multipleSendings && await _cannotImproveResults(campaign.id, userId, domainTransaction)) {
+async function _checkCanParticipateToCampaign(queryBuilder, campaign, userId) {
+  if (campaign.multipleSendings && await _cannotImproveResults(queryBuilder, campaign.id, userId)) {
     throw new ForbiddenAccess('Vous ne pouvez pas repasser la campagne');
   }
-  if (!campaign.multipleSendings && await _hasAlreadyParticipatedToCampaign(campaign.id, userId, domainTransaction)) {
+  if (!campaign.multipleSendings && await _hasAlreadyParticipatedToCampaign(queryBuilder, campaign.id, userId)) {
     throw new AlreadyExistingCampaignParticipationError(`User ${userId} has already a campaign participation with campaign ${campaign.id}`);
   }
 }
 
-async function _hasAlreadyParticipatedToCampaign(campaignId, userId, domainTransaction) {
-  const knexConn = domainTransaction.knexTransaction || knex;
-
-  const { count } = await knexConn('campaign-participations')
+async function _hasAlreadyParticipatedToCampaign(queryBuilder, campaignId, userId) {
+  const { count } = await queryBuilder('campaign-participations')
     .count('id')
     .where({ userId, campaignId })
     .first();
   return count > 0;
 }
 
-async function _cannotImproveResults(campaignId, userId, domainTransaction) {
-  const knexConn = domainTransaction.knexTransaction || knex;
-
-  const targetProfileSkillIds = await knexConn('target-profiles_skills')
+async function _cannotImproveResults(queryBuilder, campaignId, userId) {
+  const targetProfileSkillIds = await queryBuilder('target-profiles_skills')
     .select('skillId')
     .join('campaigns', 'campaigns.targetProfileId', 'target-profiles_skills.targetProfileId')
     .where({ 'campaigns.id': campaignId });
 
   const operativeTargetProfileSkillIds = await skillDatasource.findOperativeByRecordIds(targetProfileSkillIds.map(({ skillId }) => skillId));
 
-  const { count } = await knexConn('campaign-participations')
+  const { count } = await queryBuilder('campaign-participations')
     .count('id')
     .where({ userId, campaignId, isImproved: false })
     .andWhere((builder) => {
@@ -130,3 +123,5 @@ async function _cannotImproveResults(campaignId, userId, domainTransaction) {
 
   return count > 0;
 }
+
+module.exports = CampaignToJoinRepository;
